Initial Working Project
The component is able to detect the current level being played, and show
the level name and rainbow medal time for that level.

For challenges, it shows the time to beat your opponent.
For bonuses, it shows the time needed to beat the bonus.
For the credits, it shows the time you get when you just hold right
(1:01.00).